add banner to teams page. fixup naming of files
